pyproject.toml
[tool.poetry]
name = "queen-api"
version = "0.1.0"
description = ""
authors = ["Max Vieweg <maximilian.manfred.vieweg@rwth-aachen.de>"]


[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {version = "^0.17.5", extras = ["standard"]}
fire = "^0.4.0"
python-dotenv = "^0.19.2"
strawberry-graphql ={ version = "^0.99.1", extras = ["asgi"]}
pydantic = "^1.9.0"
motor = "^2.5.1"


[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest-asyncio = "^0.18.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
